+++ PoopDetector/PoopDetector.csproj
@@ -3,6 +3,8 @@
 	<PropertyGroup>
 		<TargetFrameworks>net9.0-android;net9.0-ios;net9.0-maccatalyst</TargetFrameworks>
 		<TargetFrameworks Condition="$([MSBuild]::IsOSPlatform('windows'))">$(TargetFrameworks);net9.0-windows10.0.19041.0</TargetFrameworks>
+		<TargetFrameworks Condition="!$([MSBuild]::IsOSPlatform('linux'))">$(TargetFrameworks);net9.0-ios;net9.0-maccatalyst</TargetFrameworks>
+
 		<!-- Uncomment to also build the tizen app. You will need to install tizen by following this: https://github.com/Samsung/Tizen.NET -->
 		<!-- <TargetFrameworks>$(TargetFrameworks);net8.0-tizen</TargetFrameworks> -->
 		<EnableMLUnsupportedPlatformTargetCheck>     false    </EnableMLUnsupportedPlatformTargetCheck>
@@ -12,7 +14,10 @@
 		The Mac App Store will NOT accept apps with ONLY maccatalyst-arm64 indicated;
 		either BOTH runtimes must be indicated or ONLY macatalyst-x64. -->
 		<!-- For example: <RuntimeIdentifiers>maccatalyst-x64;maccatalyst-arm64</RuntimeIdentifiers> -->
+		<!-- <RuntimeIdentifiers>ios-arm64;maccatalyst-x64;maccatalyst-arm64</RuntimeIdentifiers>  -->
+  		<!-- <RuntimeIdentifier>ios-arm64</RuntimeIdentifier> -->
 
+		<!-- <NoBindingEmbedding>false</NoBindingEmbedding> -->
 		<OutputType>Exe</OutputType>
 		<RootNamespace>PoopDetector</RootNamespace>
 		<UseMaui>true</UseMaui>
@@ -30,8 +35,8 @@
 		<ApplicationDisplayVersion>1.0</ApplicationDisplayVersion>
 		<ApplicationVersion>1</ApplicationVersion>
 
-		<SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'ios'">11.0</SupportedOSPlatformVersion>
-		<SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'maccatalyst'">13.1</SupportedOSPlatformVersion>
+		<SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'ios'">12.2</SupportedOSPlatformVersion>
+		<SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'maccatalyst'">15.0</SupportedOSPlatformVersion>
 		<SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'android'">24.0</SupportedOSPlatformVersion>
 		<SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'windows'">10.0.18362.0</SupportedOSPlatformVersion>
 		<TargetPlatformMinVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'windows'">10.0.18362.0</TargetPlatformMinVersion>
@@ -103,11 +108,11 @@
 		<PackageReference Include="Microsoft.Maui.Controls" Version="9.0.50" />
 		<PackageReference Include="Microsoft.Maui.Controls.Compatibility" Version="9.0.50" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="9.0.3" />
-		<PackageReference Include="Microsoft.ML" Version="4.0.2" />
-		<PackageReference Include="Microsoft.ML.ImageAnalytics" Version="4.0.2" />
 		<PackageReference Include="Microsoft.ML.OnnxRuntime" Version="1.21.0" />
 		<PackageReference Include="Microsoft.ML.OnnxTransformer" Version="4.0.2" />
-		<PackageReference Include="Sdcb.OpenVINO" Version="0.6.10" />
+		<PackageReference Include="Microsoft.ML.OnnxRuntime.Extensions" Version="0.14.0" />
+		<PackageReference Include="Microsoft.ML" Version="4.0.2" />
+		<PackageReference Include="Microsoft.ML.ImageAnalytics" Version="4.0.2" />
 		<PackageReference Include="SkiaSharp.Extended.UI.Maui" Version="2.0.0" />
 		<PackageReference Include="System.Drawing.Common" Version="9.0.3" />
 	</ItemGroup>
@@ -118,7 +123,6 @@
 
 	<ItemGroup>
 	  <ProjectReference Include="..\..\Camera.Maui\Camera.MAUI\Camera.MAUI.csproj" />
-	  <ProjectReference Include="..\PoopDetector.Common\PoopDetector.Common.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
